Export style into a separate file, and refactor components to be guidelines compliant
