+++ src/components/App/app.scss
@@ -1,0 +1,8 @@
+.app {
+  .results{
+    background-color: #607D8B;
+    max-width: 1000px;
+    overflow-x: hidden;
+    overflow-y: scroll;
+  }
+}
