+++ src/routes/AppRoute.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import App from 'components/App';
+import App from 'components/App/App';
 
 export default class AppRoute extends React.Component {
